requirements.txt
# Core Flask and web framework
Flask==1.1.2
Flask-Cors==3.0.8
Werkzeug>=0.15

# Mastodon integration
Mastodon.py==1.8.1
blurhash>=1.1.4
python-magic>=0.4.27

# Twitter integration
GetOldTweets3==0.0.11
tweepy==3.8.0

# Bluesky integration
websocket-client==1.6.4

# HTTP and networking
requests>=2.23.0
requests-oauthlib>=0.7.0
oauthlib>=3.0.0
urllib3>=1.25.9
certifi>=2017.4.17
PySocks>=1.5.7

# Web scraping and parsing
lxml>=4.5.1
pyquery>=1.2.10
cssselect>=1.1.0

# Utilities
python-dateutil>=2.8.1
click>=7.1.2
six>=1.14.0
chardet>=3.0.2
idna>=2.5
decorator>=4.4.2

# Template engine
Jinja2>=2.10.1
MarkupSafe>=1.1.1
itsdangerous>=0.24

# Optional: Development and notebook support (install only if needed)
# ipython>=7.14.0
# ipykernel>=5.2.1
# jupyter-client>=6.1.3
# jupyter-core>=4.6.3
# ipython-genutils>=0.2.0
# prompt-toolkit>=3.0.5
# Pygments>=2.6.1
# backcall>=0.1.0
# pexpect>=4.8.0
# pickleshare>=0.7.5
# ptyprocess>=0.6.0
# parso>=0.7.0
# jedi>=0.17.0
# wcwidth>=0.1.9
# traitlets>=4.3.3
# tornado>=6.0.4

# Note: pyzmq has compatibility issues with Python 3.13 and is not required for core functionality
# pyzmq==19.0.1  # Commented out due to Python 3.13 compatibility issues